jira-mcp/pyproject.toml
[project]
name = "jira-mcp"
version = "0.1.0"
description = "Jira MCP server for Claude AI agents - enables ticket creation and management"
readme = "README.md"
requires-python = ">=3.11,<4.0"
license = { text = "MIT" }
dependencies = [
    "fastmcp>=2.0",
    "httpx>=0.27",
    "pydantic-settings>=2.7",
]

[project.scripts]
jira-mcp = "jira_mcp.__main__:main"

[tool.poetry]
packages = [{ include = "jira_mcp", from = "src" }]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "ruff>=0.9",
]
